NIJO-351 small refactoring
