+++ src/core/routes.js
@@ -20,8 +20,9 @@ import accountInvite              from "pages/account/accept-invite";
 import userBookingManagement      from "pages/user/booking-management";
 import accountStatement           from "pages/user/account-statement";
 import userInvite                 from "pages/user/create-invite";
-import usersManagement            from "pages/user/usersManagement";
-import addNewUser                 from "pages/user/addNewUser";
+
+import usersManagement            from "pages/settings/usersManagement";
+import addNewUser                 from "pages/settings/addNewUser";
 
 import contactUsPage              from "pages/common/contact";
 import termsPage                  from "pages/common/terms";
@@ -45,8 +46,8 @@ export const routesWithHeaderAndFooter = [
     "/user/booking",
     "/user/payment-history",
     "/user/invite",
-    "/users",
-    "/users/add",
+    "/settings/users",
+    "/settings/users/add",
 ];
 export const routesWithFooter = [
     ...routesWithHeaderAndFooter,
@@ -80,8 +81,9 @@ const Routes = () => (
         <Route path="/user/booking"               component={userBookingManagement} />
         <Route path="/user/payment-history"       component={accountStatement} />
         <Route path="/user/invite"                component={userInvite} />
-        <Route path="/users/add"         component={addNewUser} />
-        <Route path="/users"                      component={usersManagement} />
+
+        <Route path="/settings/users/add"         component={addNewUser} />
+        <Route path="/settings/users"             component={usersManagement} />
 
         <Route path="/contact"                    component={contactUsPage} />
         <Route path="/terms"                      component={termsPage} />
