+++ styles/block/add-new-user.sass
@@ -2,6 +2,7 @@
     &__header
         display: flex
         justify-content: space-between
+        margin-top: 50px
         button
             font-size: 13px
             color: #A4AAB3
