+++ styles/block/user-management.sass
@@ -29,4 +29,8 @@
             &:last-child
                 border-right: 1px solid #EDEDED
     &__button-search-user
-        width: 330px
+        width: 330px !important
+    &__button-add-new-user
+        width: 180px
+        text-align: center
+        color: #fff !important
