+++ styles/blocks.sass
@@ -11,6 +11,7 @@
 @import "block/management"
 @import "block/user-management"
 @import "block/add-new-user"
+@import "block/users-pages"
 
 @import "block/account"
 @import "block/error-page"
